requirements.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2020.4.5.1
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
docker-pycreds==0.4.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.8.3
joblib==1.1.0
kiwisolver==1.3.1
lmdb==1.3.0
Markdown==3.3.6
matplotlib==3.2.1
ninja==1.11.1
numpy==1.18.4
oauthlib==3.2.0
opencv-python==4.2.0.34
pathtools==0.1.2
Pillow==7.1.2
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.4.1
sentry-sdk==1.9.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smmap==5.0.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tqdm==4.46.0
typing-extensions==4.1.1
urllib3==1.26.9
wandb==0.13.0
Werkzeug==2.0.3
zipp==3.6.0
scikit-image
torchmetrics
oyaml
pytorch_lightning
